lancaster {aggregation}	R Documentation

Lancaster method

Description

Weighted p-value aggregation.

Usage

lancaster(pvalues, weights)

Arguments

pvalues
A vector of p-values (i.e. between 0 and 1). NAs will be filtered out.

weights
A vector of weights, each associated with its respective p-value. Weights must be nonegative. NAs and negative weights will be filtered out with corresponding p-values.

Examples

lancaster(c(.1, .5), c(2, 4))

[Package aggregation version 1.0.1 Index]